+++ core/src/test/java/com/vladmihalcea/book/hpjp/hibernate/batch/OracleVersionedBatchingTest.java
@@ -143,7 +143,7 @@ public class OracleVersionedBatchingTest extends AbstractTest {
         private String title;
 
         @Version
-        private int version;
+        private short version;
 
         @OneToMany(cascade = CascadeType.ALL, mappedBy = "post",
                 orphanRemoval = true)
@@ -199,7 +199,7 @@ public class OracleVersionedBatchingTest extends AbstractTest {
         private String review;
 
         @Version
-        private int version;
+        private short version;
 
         public PostComment() {}
