+++ core/src/test/java/com/vladmihalcea/book/hpjp/hibernate/concurrency/CascadeLockUnidirectionalOneToManyTest.java
@@ -281,7 +281,7 @@ public class CascadeLockUnidirectionalOneToManyTest extends AbstractTest {
         private String body;
 
         @Version
-        private int version;
+        private short version;
 
         public Post() {}
 
@@ -353,7 +353,7 @@ public class CascadeLockUnidirectionalOneToManyTest extends AbstractTest {
         private String createdBy;
 
         @Version
-        private int version;
+        private short version;
 
         public PostDetails() {
             createdOn = new Date();
@@ -407,7 +407,7 @@ public class CascadeLockUnidirectionalOneToManyTest extends AbstractTest {
         private String review;
 
         @Version
-        private int version;
+        private short version;
 
         public PostComment() {}
