+++ core/src/test/java/com/vladmihalcea/book/hpjp/hibernate/concurrency/EntityOptimisticLockingOnBidirectionalParentOwningCollectionTest.java
@@ -29,7 +29,7 @@ public class EntityOptimisticLockingOnBidirectionalParentOwningCollectionTest
         private List<Comment> comments = new ArrayList<Comment>();
 
         @Version
-        private int version;
+        private short version;
 
         public Long getId() {
             return id;
