+++ core/src/test/java/com/vladmihalcea/book/hpjp/hibernate/concurrency/EntityOptimisticLockingOnUnidirectionalCollectionTest.java
@@ -30,7 +30,7 @@ public class EntityOptimisticLockingOnUnidirectionalCollectionTest
         private List<Comment> comments = new ArrayList<Comment>();
 
         @Version
-        private int version;
+        private short version;
 
         public Long getId() {
             return id;
