+++ core/src/test/java/com/vladmihalcea/book/hpjp/hibernate/concurrency/LockModePessimisticReadWriteIntegrationTest.java
@@ -295,7 +295,7 @@ public class LockModePessimisticReadWriteIntegrationTest extends AbstractPostgre
         private String body;
 
         @Version
-        private int version;
+        private short version;
 
         public Long getId() {
             return id;
@@ -335,7 +335,7 @@ public class LockModePessimisticReadWriteIntegrationTest extends AbstractPostgre
         private String review;
 
         @Version
-        private int version;
+        private short version;
 
         public Long getId() {
             return id;
