+++ core/src/test/java/com/vladmihalcea/book/hpjp/hibernate/concurrency/OptimisticLockingAggregateRootVersionTest.java
@@ -242,7 +242,7 @@ public class OptimisticLockingAggregateRootVersionTest extends AbstractTest {
         private String title;
 
         @Version
-        private Integer version;
+        private Short version;
 
         public Long getId() {
             return id;
