+++ core/src/test/java/com/vladmihalcea/book/hpjp/hibernate/identifier/AssignedSequenceStyleGeneratorTest.java
@@ -40,7 +40,7 @@ public class AssignedSequenceStyleGeneratorTest extends AbstractTest {
         private Long id;
 
         @Version
-        private Integer version;
+        private Short version;
 
         public Post() {
         }
