+++ core/src/test/java/com/vladmihalcea/book/hpjp/hibernate/query/dto/projection/record/PostRecordOverrideTest.java
@@ -98,7 +98,7 @@ public class PostRecordOverrideTest extends AbstractTest {
         private String updatedBy;
 
         @Version
-        private Integer version;
+        private Short version;
 
         public Long getId() {
             return id;
@@ -148,11 +148,11 @@ public class PostRecordOverrideTest extends AbstractTest {
             this.updatedBy = updatedBy;
         }
 
-        public Integer getVersion() {
+        public Short getVersion() {
             return version;
         }
 
-        public void setVersion(Integer version) {
+        public void setVersion(Short version) {
             this.version = version;
         }
     }
