+++ core/src/test/java/com/vladmihalcea/book/hpjp/hibernate/transaction/identifier/TransactionIdTest.java
@@ -70,7 +70,7 @@ public class TransactionIdTest extends AbstractTest {
         private String title;
 
         @Version
-        private int version;
+        private short version;
 
         public Long getId() {
             return id;
